requirements.txt
pandas==2.2.3
requests==2.32.3
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
browser-use==0.2.5
playwright==1.52.0
google-api-python-client==2.174.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
psycopg2-binary==2.9.10
openai==1.91.0